+++ mruby/stub_requires.rb
@@ -1,0 +1,4 @@
+module Kernel
+  def require(*args); end
+  def require_relative(*args); end
+end