+++ test/Poro.rb
@@ -33,7 +33,7 @@ describe Eventful do
   let(:machine){Machine.active.first || Machine.new}
 
   it "must have successfully extended the receiver class with Stateful methods" do
-    _(machine.class.methods.include?(:stateful_states)).must_equal true
+    _(machine.class.methods.include?(:stateful_state_machine)).must_equal true
   end
 
   it "must have successfully extended the receiver class with Eventful methods" do